Basic class prototypes/skeleton code
So here I have the classes from our UML class diagram put into some
skeleton code.

There are a few anomalies that came up as I was doing it:
- What is the "update" method for the user supposed to be actually
doing? It looks like, in the class diagram, it's returning a list of
nearby BulletinBoards, but from the name I would think it would be
making that list for the User class.
- I don't know what classes should be able to access or send updates to
the database, so some comments may be inaccurate for that kind of thing
- Did we decide how many top tags there were going to be? Maybe 3 or 5?
I left that commented out
- I renamed a couple of fields and things because I wanted to make sure
they didn't overlap with other classes we might need, or I don't know if
the Java.util class will actually work right for it